Simpl:api Studio Milestone update (#26)
* ♻️ refactor: migrate tailwindcss v3 to v4

* ✨ feat: add shadcn UI component library including alerts, buttons, cards, dialogs, dropdowns, and sidebar components


* ✨ feat: add `SearchCommand` and `ThemeSwitcher` components

* ✨ feat: add `useIsMobile` hook to detect mobile viewport size

* ✨ feat: add utility functions `cn` for className merging and `getPageNumbers` for pagination handling

* ✨ feat: add reusable `DataTable` components including pagination, toolbar, row actions, faceted filters, and view options

* ♻ refactor: integrate `ThemeProvider` and `TooltipProvider` into `Providers` component to enhance theme and UI functionality

* 🔧 chore: update `package.json` dependencies to add shadcn, tanstack table, lucide icons and zod

* ✨ feat: add simpl:api studio initial pages

* ♻️ refactor: update `globals.css` for shadcn ui components

* ✨ feat: add `PublishStatus` enum and integrate it into relevant interfaces

* ✨ feat: add `ROUTES` constant for defining application paths

* ✨ feat: add sign in and sign up zod schemas

* ✨ feat: add user-related schemas including `UserCreateSchema`, `UserUpdateSchema`, `UserDeleteSchema`, and validation for profile images

* ✨ feat: enhance DataTable components with user and entry type status handling

* ✨ feat: extend interfaces with `UserStatus` enum, additional `User` fields, and enhanced `PermissionGroup` structure

* ✨ feat: integrate `AvatarGroup` into `PermissionGroupDataTable` and improve name column with icon display

* ✨ feat: add shadcn avatar and form components

* ✨ feat: add extended schemas for entries, entry types, permissions, and privileges

* ✨ feat: add type definitions for entries, users, entry types, and permission groups using Zod schemas

* 🔧 chore: add linting rule to allow children prop for tanstack form

* ✨ feat: add new app management UI `simpl-api studio` and its new features
♻️ refactor: adapt `api/v1` and other backend functions to new `simpl-api studio`

* 🐛 fix: fix TS errors and logic bugs, remove unused code

* 🔥 remove: delete outdated API, tests, and Cypress scenarios

* 🐛 fix: resolve TS errors, improve error handling, and enhance type safety

* ♻️ refactor: centralize API key validation with `isValidApiKey` utility and fix TS errors.

* 📄 docs: change license to GNU AGPL v3

* ✨ feat: integrate Redis for atomic rate limiting and improve request validation

* ✨ feat: implement system API key validation and permission checks across all endpoints

* ♻️ refactor: clean up unused `finally` blocks across controllers

* 🔧 chore: update project metadata and improve tooling

- Add project metadata (`name`, `author`, `version`) in `package.json`.
- Introduce new npm scripts for linting, formatting, and type checking.
- Add missing dependency `ioredis`.
- Update TypeScript configuration and exclude unnecessary files.

* 📄 docs: add SECURITY.md

* 📄 docs: add DEVELOPMENT.md

* 📄 docs: add CODE-OVERVIEW.md

* 📄 docs: add CONTRIBUTING.md

* 📄 docs: add AI related docs AI.md and AGENTS.md

* 📄 docs: add Contributor License Agreement (CLA)

* 📄 docs: update README.md for v0.5.0 (new Simpl:API Studio update)

* 🐛 fix: update required permissions to `read`

* 📄 doc: update project overview to make it consistent across all documents

* ♻️ refactor: fix eslint errors

* ♻️ refactor: remove unnecessary code

* ♻️ refactor: replace removed icons from the icon library

* ♻️ refactor: update global `redis` declaration to use explicit interface

* 📄 docs: add JsDoc comments for Redis options and client configuration

* ♻️ refactor: update Cypress ESLint plugin import path

* 🔧 chore: update eslint config

* 🔧 chore: add `__mocks__` to exclude list in tsconfig.json

* 🔧 chore: update package versions

* 🔧 chore: add .swc/ to ignored list in .gitignore

* 📄 docs: update setup process

* ✨ feat: add `checkPermissionApi` utility function for multiple permission validation

* ♻️ **refactor:** replace `hasPermissionApi` with `checkPermissionApi` across API routes for enhanced permission validation

* :memo: docs: update README.md

* ♻️ **refactor:** update dropdown links

* ♻️ **refactor:** update dropdown links

* ✨ **feat:** add `buildPatternRegExp` utility for safer regex handling in entry validation

* ♻️ refactor: fix eslint errors

* ♻️ refactor: fix eslint errors

* ♻️ refactor: fix eslint errors

* ♻️ refactor: fix eslint errors

* 🔧 chore: update nodemailer version

* ♻️ refactor: fix eslint errors

---------

+++ controllers/api-key.controller.ts
@@ -1,23 +1,22 @@
-import { connectDB } from "../lib/mongodb"
-import { Collection, DeleteResult, Filter, InsertOneResult, MongoClient, ObjectId } from "mongodb"
+import { connectDB } from "@/lib/mongodb"
+import { Collection, DeleteResult, InsertOneResult, MongoClient, ObjectId } from "mongodb"
 
 //Interface
-import { ApiKey } from "../interfaces"
+import { ApiKey } from "@/interfaces"
 
 //===============================================
 
 export class apiKeyController {
-  apiKey: ApiKey
+  apiKey: Partial<ApiKey>
 
-  constructor(apiKey: ApiKey) {
+  constructor(apiKey: Partial<ApiKey>) {
     this.apiKey = apiKey
   }
 
-  async findKey() {
-    let dbCollection: Collection
+  async findKey(): Promise<ApiKey[] | { message: string }[] | undefined> {
+    let dbCollection: Collection<ApiKey>
     let isConnected = false
-    let client: MongoClient
-
+    let client: MongoClient | undefined
     try {
       try {
         client = await connectDB()
@@ -25,11 +24,11 @@ export class apiKeyController {
       } catch (e) {
         console.log(e)
       }
-      if (isConnected) {
-        let findResult: Filter<object>
+      if (client && isConnected) {
+        let findResult: ApiKey[] | undefined
         try {
-          dbCollection = client.db(process.env.DB_NAME).collection("api_keys") as Collection
-          findResult = await dbCollection.find({ key: this.apiKey.key }).toArray()
+          dbCollection = client.db(process.env.DB_NAME).collection("api_keys") as Collection<ApiKey>
+          findResult = (await dbCollection.find({ key: this.apiKey.key }).toArray()) as ApiKey[]
         } catch (e) {
           console.log(e)
         }
@@ -39,17 +38,13 @@ export class apiKeyController {
       }
     } catch (e) {
       console.error(e)
-    } finally {
-      if (client?.close && typeof client.close === "function") {
-        await client.close()
-      }
     }
   }
 
   async create() {
-    let dbCollection: Collection
+    let dbCollection: Collection<any>
     let isConnected = false
-    let client: MongoClient
+    let client: MongoClient | undefined
 
     try {
       try {
@@ -58,8 +53,8 @@ export class apiKeyController {
       } catch (e) {
         console.log(e)
       }
-      if (isConnected) {
-        let insertResult: InsertOneResult
+      if (client && isConnected) {
+        let insertResult: InsertOneResult | undefined
         try {
           dbCollection = client.db(process.env.DB_NAME).collection("api_keys")
           insertResult = await dbCollection.insertOne(this.apiKey)
@@ -67,7 +62,7 @@ export class apiKeyController {
           console.log(e)
         }
 
-        if (insertResult.insertedId) {
+        if (insertResult && insertResult.insertedId) {
           return { status: "success", message: "API Key has been generated.", keyId: insertResult.insertedId }
         } else {
           return { status: "failed", message: "Failed to create the api key." }
@@ -77,16 +72,12 @@ export class apiKeyController {
       }
     } catch (e) {
       console.error(e)
-    } finally {
-      if (client?.close && typeof client.close === "function") {
-        await client.close()
-      }
     }
   }
 
   async delete(id: string) {
-    let client: MongoClient
-    let dbCollection: Collection
+    let client: MongoClient | undefined
+    let dbCollection: Collection<any>
     let isConnected = false
 
     try {
@@ -96,15 +87,15 @@ export class apiKeyController {
       } catch (e) {
         console.log(e)
       }
-      if (isConnected) {
-        let deleteResult: DeleteResult
+      if (client && isConnected) {
+        let deleteResult: DeleteResult | undefined
         try {
           dbCollection = client.db(process.env.DB_NAME).collection("api_keys")
           deleteResult = await dbCollection.deleteOne({ _id: new ObjectId(id) })
         } catch (e) {
-          console.log("deleteResult:", deleteResult, "Error:", e) //TODO: better error logging & displaying
+          console.error("Failed to delete API key:", e) //TODO: better error logging & displaying
         }
-        if (deleteResult.deletedCount === 1) {
+        if (deleteResult && deleteResult.deletedCount === 1) {
           return { status: "success", message: "API Key has been removed." }
         } else {
           return { status: "failed", message: "Failed to delete the api key." }
@@ -113,11 +104,7 @@ export class apiKeyController {
         return [{ message: "Database connection is NOT established" }]
       }
     } catch (e) {
-      console.log(e) //TODO: better error logging & displaying
-    } finally {
-      if (client?.close && typeof client.close === "function") {
-        await client.close()
-      }
+      console.error("An error occurred while deleting the API key:", e) //TODO: better error logging & displaying
     }
   }
 }
