+++ controllers/entry.controller.ts
@@ -1,9 +1,9 @@
 //Database
-import { connectDB } from "../lib/mongodb"
+import { connectDB } from "@/lib/mongodb"
 import { Collection, DeleteResult, InsertOneResult, MongoClient, ObjectId, UpdateResult } from "mongodb"
 
 //Interface
-import { Entry } from "../interfaces"
+import { Entry } from "@/interfaces/entry"
 
 export class EntryController {
   entry: Entry
@@ -14,8 +14,8 @@ export class EntryController {
   }
 
   async create() {
-    let client: MongoClient
-    let dbCollection: Collection
+    let client: MongoClient | undefined
+    let dbCollection: Collection<any>
     let isConnected = false
 
     try {
@@ -24,7 +24,7 @@ export class EntryController {
     } catch (e) {
       console.log(e)
     }
-    if (isConnected) {
+    if (client && isConnected) {
       let insertResult: InsertOneResult
       try {
         dbCollection = client.db(process.env.DB_NAME).collection("entries")
@@ -44,10 +44,6 @@ export class EntryController {
       } catch (e) {
         console.log(e)
         return { status: "failed", message: "Failed to create the entry." }
-      } finally {
-        if (client?.close && typeof client.close === "function") {
-          await client.close()
-        }
       }
     } else {
       return [{ message: "Database connection is NOT established" }]
@@ -55,8 +51,8 @@ export class EntryController {
   }
 
   async update(id: string) {
-    let client: MongoClient
-    let dbCollection: Collection
+    let client: MongoClient | undefined
+    let dbCollection: Collection<any>
     let isConnected = false
 
     try {
@@ -65,21 +61,17 @@ export class EntryController {
     } catch (e) {
       console.log(e)
     }
-    if (isConnected) {
-      let updateResult: UpdateResult
+    if (client && isConnected) {
+      let updateResult: UpdateResult | undefined
       try {
         dbCollection = client.db(process.env.DB_NAME).collection("entries")
         updateResult = await dbCollection.updateOne({ _id: new ObjectId(id) }, { $set: this.entry }, { upsert: false })
       } catch (e) {
         console.log(e)
-      } finally {
-        if (client?.close && typeof client.close === "function") {
-          await client.close()
-        }
       }
-      if (updateResult.modifiedCount === 1) {
+      if (updateResult && updateResult.modifiedCount === 1) {
         return { status: "success", message: "Entry has been updated." }
-      } else if (updateResult.matchedCount === 1 && updateResult.modifiedCount === 0) {
+      } else if (updateResult && updateResult.matchedCount === 1 && updateResult.modifiedCount === 0) {
         return { status: "failed", message: "You didn't make any change." }
       } else {
         return { status: "failed", message: "Failed to update the entry." }
@@ -90,8 +82,8 @@ export class EntryController {
   }
 
   async delete(id: string) {
-    let client: MongoClient
-    let dbCollection: Collection
+    let client: MongoClient | undefined
+    let dbCollection: Collection<any>
     let isConnected: boolean = false
 
     try {
@@ -100,19 +92,15 @@ export class EntryController {
     } catch (e) {
       console.log(e)
     }
-    if (isConnected) {
-      let deleteResult: DeleteResult
+    if (client && isConnected) {
+      let deleteResult: DeleteResult | undefined
       try {
         dbCollection = client.db(process.env.DB_NAME).collection("entries")
         deleteResult = await dbCollection.deleteOne({ _id: new ObjectId(id) })
       } catch (e) {
         console.log(e)
-      } finally {
-        if (client?.close && typeof client.close === "function") {
-          await client.close()
-        }
       }
-      if (deleteResult.deletedCount === 1) {
+      if (deleteResult && deleteResult.deletedCount === 1) {
         return { status: "success", message: "Entry has been deleted." }
       } else {
         return { status: "failed", message: "Failed to delete the entry." }
